chore(table): 代码调整
