+++ .changeset/stupid-toes-pull.md
@@ -2,4 +2,4 @@
 "@hi-ui/hiui": patch
 ---
 
-Table feat: setting 设置面板对外暴露打开和关闭方法
+Table feat: 对外暴露 setting 面板，用户可以自己控制显示和隐藏
