+++ packages/ui/table/src/index.ts
@@ -3,4 +3,6 @@ import './styles/index.scss'
 export * from './Table'
 export { Table as default } from './Table'
 
+export { TableSettingMenuDrawer } from './TableSettingMenuDrawer'
+
 export * from './types'
