+++ packages/ui/table/src/types.ts
@@ -291,8 +291,3 @@ export interface TableRowRequiredProps {
 }
 
 export type TableRowRecord = Record<string, any>
-
-export interface SettingMenuHelper {
-  open: () => void
-  close: () => void
-}
