+++ packages/ui/table/stories/index.stories.tsx
@@ -30,6 +30,7 @@ export * from './merged-cell.stories'
 export * from './pagination.stories'
 export * from './data-source.stories'
 export * from './setting.stories'
+export * from './custom-setting.stories'
 export * from './loading.stories'
 export * from './empty.stories'
 export * from './virtual.stories'
